requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
PyMySQL==1.1.1
python-dotenv==1.0.1
bcrypt==4.1.3
cryptography==42.0.8
opencv-python==4.8.1.78
baidu-aip==4.16.11
setuptools==70.3.0
numpy==1.26.4
Pillow==10.3.0
